# openapi/music-library-management.yaml
openapi: 3.0.0
info:
  title: Music API
  description: An API for managing music tracks and playlists
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /tracks:
    post:
      summary: Create a new music track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/track_input'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/track'
    get:
      summary: Get all music tracks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/track'
  /tracks/{id}:
    get:
      summary: Get a music track by ID
      parameters:
        - name: id
          in: path
          description: ID of the track to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/track'
    put:
      summary: Update a music track
      parameters:
        - name: id
          in: path
          description: ID of the track to update
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/track_input'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/track'
    delete:
      summary: Delete a music track
      parameters:
        - name: id
          in: path
          description: ID of the track to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /playlists:
    # CRUD operations for playlists
  /search:
    # Search functionality
components:
  schemas:
    track:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        genre:
          type: string
        release_year:
          type: integer
        duration:
          type: integer
      required:
        - title
        - artist
        - album
    track_input:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        genre:
          type: string
        release_year:
          type: integer
        duration:
          type: integer